requirements.txt
torch
torchvision
numpy
opencv-python
matplotlib
streamlit
streamlit-drawable-canvas
Pillow
